refactor: Vector2F uses float and Vector2 uses int
Spritesheet coordinates and size are always integers so they can now be used without forced conversion to float by the old Vector2.

+++ Pac-Man-DeLuxe/Pac-Man-DeLuxe/GameObject.cpp
@@ -8,7 +8,7 @@ GameObject::GameObject(float render_x, float render_y, float width, float height
 	this->id_ = ++current_id_;
 }
 
-GameObject::GameObject(float render_x, float render_y, float width, float height, float spritesheet_x, float spritesheet_y, float sprite_width, float sprite_height)
+GameObject::GameObject(float render_x, float render_y, float width, float height, int spritesheet_x, int spritesheet_y, int sprite_width, int sprite_height)
 	: render_position_(render_x, render_y), render_size_(width, height), spritesheet_position_(spritesheet_x, spritesheet_y), sprite_size_(sprite_width, sprite_height)
 {
 	this->id_ = ++current_id_;
