+++ Pac-Man-DeLuxe/Pac-Man-DeLuxe/Vector2.h
@@ -1,9 +1,9 @@
 #pragma once
 
 struct Vector2 {
-	float x_;
-	float y_;
+	int x_;
+	int y_;
 
 	Vector2();
-	Vector2(float x, float y);
+	Vector2(int x, int y);
 };
